+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/table/DefaultFilteredTableModel.java
@@ -433,7 +433,7 @@ final class DefaultFilteredTableModel<R, C> extends AbstractTableModel implement
 			}
 		});
 		addTableModelListener(removeSelectionListener);
-		filterModel.addChangeListener(this::filterItems);
+		filterModel.changeObserver().addListener(this::filterItems);
 		sortModel.addSortingChangedListener(columnIdentifier -> sortItems());
 	}