+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilteredTable.java
@@ -701,7 +701,7 @@ public final class FilteredTable<R, C> extends JTable {
 
 	private void bindEvents() {
 		tableModel.selectionModel().selectedIndexesObserver().addDataListener(new ScrollToSelectedListener());
-		tableModel.filterModel().addChangeListener(getTableHeader()::repaint);
+		tableModel.filterModel().changeObserver().addListener(getTableHeader()::repaint);
 		tableModel.searchModel().addCurrentResultListener(rowColumn -> repaint());
 		tableModel.sortModel().addSortingChangedListener(columnIdentifier -> getTableHeader().repaint());
 		addMouseListener(new FilteredTableMouseListener());